+++ common/app/menu/menu_item/menu_item_exit/menu_item_exit.cpp
@@ -3,5 +3,5 @@
 namespace common {
 std::string_view MenuItemExit::getName() const { return "exit"; }
 
-int MenuItemExit::getArgsCount() { return 0; }
+size_t MenuItemExit::getArgsCount() { return 0; }
 } // namespace common
